reagrranged a bunch of things and made alot of changes. Really messed up my commits. Turned out I typed const bcryptjs = instead of const bcrypt = required('bcryptjs') and that was causing my login not to work
