+++ auth/restricted-middleware.js
@@ -1,3 +1,7 @@
+const bcrypt = require('bcryptjs');
+
+const Users = require('../users/users-model.js');
+
 module.exports = function restricted(req, res, next) {
     const { username, password } = req.headers;
