+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "and": "0.0.3",
     "bcryptjs": "^2.4.3",
+    "cors": "^2.8.5",
     "dotenv": "^8.1.0",
     "express": "^4.17.1",
     "helmet": "^3.20.1",
